backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
anthropic==0.7.4
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.0
anyio==3.7.1